src/requirements/base.txt
django==4.2.11
django-environ==0.10.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-filter==22.1
django-autoslug==1.9.8
django-taggit==4.0.0
PyJWT==2.8.0 
drf-yasg==1.21.5
pillow==9.4.0
redis==4.5.1
celery==5.2.7
flower==1.2.0